% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_segmentation.R
\name{binary_segmentation}
\alias{binary_segmentation}
\title{Binary segmentation}
\usage{
binary_segmentation(y, threshold = NULL, maxiter = NULL)
}
\arguments{
\item{y}{A numeric vector.}

\item{threshold}{Threshold for determining changepoint candidates; defaults to sqrt(2*log(n)*var(y)).}

\item{maxiter}{Maximum number of changepoints to find; defaults to length(y) - 1}
}
\value{
A list.
}
\description{
Binary segmentation algorithm for detecting changepoints, in the change in mean model.
The algorithm terminates when either (i) the maximum number of changepoints have been found; or
(ii) there are no points remaining where the CUSUM statistic is above the threshold.
}
\examples{
set.seed(10)
y <- rnorm(100) + c(rep(1,20), rep(-1,30), rep(1,50))
binary_segmentation(y)

}
