% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_segmentation_psi.R
\name{binary_segmentation_psi}
\alias{binary_segmentation_psi}
\title{Post-selection inference for binary segmentation}
\usage{
binary_segmentation_psi(
  y,
  results = NULL,
  nus = NULL,
  threshold = NULL,
  maxiter = NULL,
  eps0 = 0.01,
  sigma2 = NULL,
  h = NULL,
  first_cp_only = FALSE,
  num_pvals = NULL
)
}
\arguments{
\item{y}{A numeric vector of data.}

\item{results}{Output of binary_segmentation function; can be NULL.}

\item{nus}{List of nu's for each detected changepoint.}

\item{threshold}{Minimum threshold for CUSUM statistic for changepoint detection; ignored if results specified.}

\item{maxiter}{Max. number of changepoints to find; ignored if results specified; otherwise defaults to n-1.}

\item{eps0}{Hyperparameter for calculating S.}

\item{sigma2}{Variance of y. If unknown, it can be estimated within the function.}

\item{h}{Window size. If NULL, the window is taken to be ...}

\item{first_cp_only}{Logical. Should be TRUE if the condition is that the changepoint of interest is in the model;
FALSE if the condition is that all changepoints are the same.}

\item{num_pvals}{Integer. Maximum number of p-values to calculate; defaults to n-1 if NULL.}
}
\value{
A list containing: the vector of changepoints and the vector of p-values.
}
\description{
Calculate p-values associated with detected changepoints for binary segmentation;
uses method of Jewell et al.
}
\details{
...
}
\examples{
set.seed(100)
y <- rnorm(200) + c(rep(1,50), rep(0,100), rep(1,50))
results <- binary_segmentation(y)
pvals <- binary_segmentation_psi(y, results, sigma2=1)
print(pvals)

}
