% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_interval_not.R
\name{calculate_interval_not}
\alias{calculate_interval_not}
\title{Calculate interval for narrowest over threshold}
\usage{
calculate_interval_not(y, nu, results, nu2 = NULL, nuTy = NULL)
}
\arguments{
\item{y}{Numeric vector of data.}

\item{nu}{Numeric vector.}

\item{results}{Output of \code{narrowest_over_threshold}.}

\item{nu2}{Value of \eqn{||\nu||_2^2}.}

\item{nuTy}{Value of \eqn{\nu^T y}.}
}
\value{
A 2-dimensional vector.
}
\description{
Find values of phi which satisfy the required inequalities so that applying narrowest over threshold to \eqn{y'(\phi)}
returns \code{(b, d)}.
}
\details{
Used inside \code{calculate_S} if \code{method = "not"}.
}
\examples{
set.seed(100)
y <- rnorm(100) + c(rep(1,50), rep(-1,50))
results <- narrowest_over_threshold(y, lambda=4, N=50)
b <- results$results$b
h <- 10
nu <- c(rep(0, b[1]-h), rep(1/h, h), rep(-1/h, h), rep(0, length(y)-b[1]-h))
calculate_interval_not(y, nu, results=results)

}
