% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_interval_wbs.R
\name{calculate_interval_wbs}
\alias{calculate_interval_wbs}
\title{Calculate interval for wild binary segmentation}
\usage{
calculate_interval_wbs(
  y,
  nu,
  results = NULL,
  b = NULL,
  d = NULL,
  s = NULL,
  e = NULL,
  rand_ints = NULL,
  nu2 = NULL,
  nuTy = NULL,
  threshold = NULL,
  n.cp = NULL
)
}
\arguments{
\item{y}{vector of data (ignored if cs & nuTy are supplied)}

\item{nu}{...}

\item{results}{Output of \code{wild_binary_segmentation}.}

\item{b}{Numeric vector of changepoints.}

\item{d}{Numeric vector containing directions of changepoints; all entries should be either +1 or -1.}

\item{s}{...}

\item{e}{...}

\item{rand_ints}{...}

\item{nu2}{value of ||nu||_2^2}

\item{nuTy}{value of nu^T y}

\item{threshold}{Changepoint threshold used in wild binary segmentation algorithm.}

\item{n.cp}{Integer: maximum number of changepoints to detect.}
}
\value{
A 2-dimensional vector
}
\description{
Find values of phi which satisfy the required inequalities so that the wild binary segmentation algorithm returns
\code{(b, d)}.
Option to only consider part of \code{(b, d)}, e.g. if we just want b1 & d1 to be specified, & don't care about
later values
}
\details{
(possibly outdated)
At present this ignores the threshold and just matches the number of CPs found (currently updating)
If threshold is also supplied, this will also be used.
Otherwise, if threshold is supplied (but not n.cp), the interval returned will be that for which the complete
b and d match those given, when BS is given this threshold. If threshold is not supplied, it will be assumed
that BS is run with a fixed number of iterations. If both n.cp and threshold are supplied, it will be assumed that
n.cp is the maximum number of iterations, but the threshold will also be used as a minimum for C(t).
}
\examples{
x <- 0

}
