% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pvals.R
\name{calculate_pvals}
\alias{calculate_pvals}
\title{Calculate p-values for change in mean model}
\usage{
calculate_pvals(
  y,
  results = NULL,
  method = "bs",
  N = 100,
  nus = NULL,
  threshold = NULL,
  maxiter = NULL,
  h = 2,
  sigma2 = 1,
  eps0 = 0.01,
  include_original = FALSE,
  num_pvals = NULL,
  random_samples = NULL,
  num_rand_samples = NULL,
  seeded = FALSE,
  decay = NULL,
  return_probs = FALSE
)
}
\arguments{
\item{y}{Numeric vector of data.}

\item{results}{Output of changepoint algorithm (either \code{binary_segmentation}, \code{wild_binary_segmentation}, or
\code{narrowest_over_threshold}.}

\item{method}{Method used to generate changepoints. Options: \code{"bs"} for binary segmentation,
\code{"wbs"} for wild or seeded binary segmentation, \code{"not"} for narrowest over threshold.}

\item{N}{Positive integer. Number of random \eqn{\psi}'s to generate.}

\item{nus}{List of nu's for each changepoint.}

\item{threshold}{Minimum threshold for changepoint detection. Ignored if \code{results} is specified.}

\item{maxiter}{Positive integer. Maximum number of changepoints to find in changepoint algorithm. Ignored if
\code{results} is specified}

\item{h}{Positive integer (>=2) or \code{NULL}. Window size. If \code{NULL} then the changepoints either side of the
changepoint of interest are used to define the window.}

\item{sigma2}{Variance of \code{y}.}

\item{eps0}{Hyperparameter for calculating S.}

\item{include_original}{Logical. Whether to include the \eqn{\psi} value corresponding to the observed data in place of
one of the random samples.}

\item{num_pvals}{Integer or NA. Maximum number of p-values to calculate; if set to NA, it will default to \code{length(y) - 1}.
If the number of changepoints detected by the binary segmentation algorithm is less than or equal to \code{num_pvals}, then
p-values will be calculated for all changepoints. If \code{num_pvals} is less than the number of changepoints detected,
p-values will only be calculated for the first \code{num_pvals} changepoints, in order of detection.}

\item{random_samples}{Matrix containing random intervals. For wild binary segmentation and narrowest over
threshold only.}

\item{num_rand_samples}{Number of random intervals to use for wild binary segmentation or narrowest over threshold.
Ignored if \code{random_samples} is specified, \code{method = "bs"}, or \code{seeded = TRUE}.}

\item{seeded}{Logical. If \code{TRUE}, use seeded binary segmentation rather than wild binary segmentation. Only used if
\code{method = "wbs"}.}

\item{decay}{Decay parameter for seeded binary segmentation. Only used if \code{method = "wbs"} and \code{seeded = TRUE}.}

\item{return_probs}{Logical. If \code{TRUE}, the values of \eqn{Pr(\phi \in S & |\phi| > |\phi_{obs}|)} and \eqn{Pr(\phi \in S)}
for each \eqn{\psi} will be included in the output.}
}
\value{
A vector of p-values
}
\description{
Calculate p-values for change in mean model
}
\details{
Given a changepoint of interest \eqn{\tau_j}, there are two options for the null hypothesis:
\itemize{
\item There are no changepoints within a window size \eqn{h} of \eqn{\tau_j}. If \code{nus = NULL} but \code{h} is supplied,
then the value of \eqn{\nu} for each changepoint will be calculated using this assumption.
\item There are no other changepoints between \eqn{\tau_{j-1}} and \eqn{\tau_{j+1}}. If \eqn{\tau_j} is the first
changepoint, then \eqn{\tau_{j-1}} is taken to be 0; if it is the last changepoint, \eqn{\tau_{j+1}} is taken to be
\code{length(y)}. If \code{nus = NULL} and \code{h = NULL}, then the value of \eqn{\nu} for each changepoint will be calculated
using this assumption.
}
Alternatively \code{nus} can be specified manually.
}
\examples{
# to do

}
