% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_S.R
\name{calculate_S}
\alias{calculate_S}
\title{Calculate S}
\usage{
calculate_S(
  y,
  nu,
  results = NULL,
  b = NULL,
  d = NULL,
  threshold = NULL,
  maxiter = NULL,
  nu2 = NULL,
  nuTy = NULL,
  eps0 = 0.01,
  first_cp_only = FALSE,
  method = "bs",
  rand_ints = NULL,
  seeded = FALSE,
  decay = NULL
)
}
\arguments{
\item{y}{Numeric vector of data.}

\item{nu}{Numeric vector.}

\item{results}{Output of changepoint algorithm (either \code{binary_segmentation}, \code{wild_binary_segmentation}, or
\code{narrowest_over_threshold}.}

\item{b}{Vector of changepoints; ignored if \code{results} specified.}

\item{d}{Vector containing directions of changepoints (+1 or -1); ignored if \code{results} specified.}

\item{threshold}{Threshold for detecting changepoints; ignored if \code{results} specified.}

\item{maxiter}{Maximum number of changepoints to detect; ignored if \code{results} specified.}

\item{nu2}{Value of \eqn{||\nu||_2^2}; optional.}

\item{nuTy}{Value of \eqn{nu^T y}; optional.}

\item{eps0}{Hyperparameter.}

\item{first_cp_only}{Logical. If \code{TRUE}, condition on the fact that the changepoint of interest is in the model;
if \code{FALSE}, condition on all changepoints. Defaults to \code{FALSE}.}

\item{method}{Character. One of \code{"bs"} (binary segmentation), \code{"wbs"} (wild binary segmentation), or \code{"not"} (narrowest
over threshold). Defaults to \code{"bs"}.}

\item{rand_ints}{Matrix containing random intervals for changepoint algorithm. Ignored if \code{results} specified
or \code{method = "bs"}.}

\item{seeded}{Logical. For \code{method = "wbs"} only, whether to use seeded binary segmentation.}

\item{decay}{Decay parameter for seeded binary segmentation. Only used if \code{method = "wbs"} and \code{seeded = TRUE}.}

\item{num_pvals}{Integer. Maximum number of p-values to calculate; defaults to \code{length(y) - 1}.}
}
\value{
A dataframe containing intervals with the changepoints obtained when \eqn{\phi} is in each interval.
}
\description{
A function for calculating the set of interest S, when the changepoint algorithm used is one of binary segmentation,
wild binary segmentation, or narrowest over threshold.
}
\examples{
set.seed(100)
y <- rnorm(100) + c(rep(1,50), rep(-1,50))
results <- binary_segmentation(y)
print(results$changepoints)
h <- 10
nu <- c(rep(0, results$changepoints[1] - h), rep(1/h, h), rep(-1/h, h), rep(0, length(y) - results$changepoints[1] - h))
calculate_S(y, nu, results, method="bs", first_cp_only=TRUE)

}
