% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_S_all_methods.R
\name{calculate_S_all_methods}
\alias{calculate_S_all_methods}
\title{Calculate S}
\usage{
calculate_S_all_methods(
  y,
  nu,
  results = NULL,
  b = NULL,
  d = NULL,
  nu2 = NULL,
  nuTy = NULL,
  threshold = NULL,
  maxiter = NULL,
  eps0 = 0.01,
  first_cp_only = FALSE,
  method = "bs",
  rand_ints = NULL,
  seeded = FALSE,
  decay = NULL
)
}
\arguments{
\item{y}{Numeric vector of data.}

\item{nu}{...}

\item{results}{Output of changepoint algorithm.}

\item{b}{Vector of changepoints; ignored if results specified.}

\item{d}{Vector containing directions of changepoints (+1 or -1); ignored if results specified.}

\item{nu2}{Scalar; value of ||nu||_2^2; optional.}

\item{nuTy}{Value of nu^T y.}

\item{threshold}{Threshold for detecting changepoints; ignored if results specified.}

\item{maxiter}{Maximum number of changepoints to detect; ignored if results specified.}

\item{eps0}{...}

\item{first_cp_only}{...}

\item{method}{One of "bs", "wbs", or "not": binary segmentation, wild/seeded binary segmentation, narrowest
over threshold. Defaults to binary segmentation.}

\item{rand_ints}{Matrix containing random intervals for changepoint algorithm. Ignored if results specified
or method = "bs".}

\item{seeded}{For method = "wbs" only, whether to use seeded binary segmentation. (Logical.)}

\item{decay}{Decay parameter for seeded binary segmentation. Only used if method = "wbs" and seeded = TRUE.}
}
\value{
A dataframe containing intervals with the changepoints obtained when phi is in each interval.
}
\description{
A function for calculating the set of interest S.
}
\details{
...
}
\examples{
# ...
}
