% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoints_psi.R
\name{changepoints_psi}
\alias{changepoints_psi}
\title{Post-selection inference for changepoints}
\usage{
changepoints_psi(
  y,
  results = NULL,
  nu = NULL,
  threshold = NULL,
  maxiter = NULL,
  eps0 = 0.01,
  sigma2 = NULL,
  h = NULL,
  first_cp_only = FALSE,
  method = "bs",
  num_rand_ints = NULL,
  rand_ints = NULL,
  return_S = TRUE
)
}
\arguments{
\item{y}{Numeric vector of data.}

\item{results}{Output of changepoint algorithm (either \code{binary_segmentation}, \code{wild_binary_segmentation}, or
\code{narrowest_over_threshold}).}

\item{nu}{nu}

\item{threshold}{Changepoint detection threshold.}

\item{maxiter}{Integer. Maximum number of changepoints for algorithm to detect.}

\item{eps0}{...}

\item{sigma2}{Variance of \code{y}.}

\item{h}{Window size.}

\item{first_cp_only}{...}

\item{method}{Character string; one of \code{"bs"} (binary segmentation), \code{"wbs"} (wild binary segmentation),
or \code{"not"} (narrowest over threshold).}

\item{num_rand_ints}{Number of random intervals for changepoint algorithm (not for \code{method = "bs"}).}

\item{rand_ints}{Random intervals for changepoint algorithm.}

\item{return_S}{Logical. Whether to return the set \code{S}.}
}
\value{
...
}
\description{
Apply a changepoint algorithm (binary segmentation, wild binary segmentation, seeded binary segmentation, or narrowest
over threshold) to a vector of data, where the model is piecewise constant mean. Compute the p-value associated with the
first changepoint using method of Jewell et al.
}
\details{
Either \code{results} or \code{threshold} and \code{maxiter} should be given.
}
\examples{
# to do

}
