% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusum_phi_vec.R
\name{cusum_phi_vec}
\alias{cusum_phi_vec}
\title{Calculate vector of CUSUM statistics for data y, in terms of phi}
\usage{
cusum_phi_vec(y, nu, nu2 = NULL, nuTy = NULL, s = 1, e = length(y))
}
\arguments{
\item{y}{Numeric vector of data.}

\item{nu}{Numeric vector.}

\item{nu2}{Numeric; value of the squared 2-norm of \code{nu}.}

\item{nuTy}{Value of nu^T y.}

\item{s}{Starting point for calculating CUSUM statistics, defaults to 1.}

\item{e}{Ending point for calculating CUSUM statistics, degaults to \code{length(y)}.}
}
\value{
Returns \code{length(y)} x 2 matrix such that
\code{cusum_phi_vec(y, nu)^T \%*\% c(1, phi) = cusum(y_phi(y, nu, phi))}.
}
\description{
Used inside \code{calculate_pvals}.
}
\examples{
set.seed(100)
y <- rnorm(100) + c(rep(1,40), rep(-1,60))
results <- binary_segmentation(y)
b <- results$changepoints[1]
nu <- c(rep(1/b, b), rep(-1/(100-b), 100-b))
cusum_phi_vec(y, nu)

}
