% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mad.R
\name{mad}
\alias{mad}
\title{Estimate variance by median absolute deviation}
\usage{
mad(y)
}
\arguments{
\item{y}{Numeric vector of data.}
}
\value{
Scalar: estimated variance of \code{y}.
}
\description{
Estimate variance of a vector using median absolute deviation, under the assumption that the data follows a
Normal distribution.
}
\examples{
y <- rnorm(200, sd=2) + c(rep(1, 100), rep(-1, 100))
mad(y)

}
