% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narrowest_over_threshold.R
\name{narrowest_over_threshold}
\alias{narrowest_over_threshold}
\title{Narrowest over threshold for change in mean.}
\usage{
narrowest_over_threshold(y, lambda, rand_ints = NULL, N = 1000, max_cps = NULL)
}
\arguments{
\item{y}{A numeric vector of data.}

\item{lambda}{Threshold for CUSUM statistic; a scalar.}

\item{rand_ints}{N x 2 matrix containing random intervals; optional.}

\item{N}{Number of random intervals; defaults to 1000. Ignored if \code{rand_ints} is specified.}

\item{max_cps}{Maximum number of changepoints to return; defaults to \code{length(y) - 1} if \code{NULL}.}
}
\value{
List.

\code{results}. Data frame.

\code{changepoints}. Vector of changepoints detected.

\code{rand_ints}. N x 2 matrix containing random intervals used in the algorithm.

\code{threshold}. Value of the threshold used in the algorithm.

\code{maxiter}. Maximum number of changepoints to detect.
}
\description{
Apply the narrowest over threshold changepoint algorithm to a vector of data. The model is the piecewise constant mean model.
}
\examples{
y <- rnorm(200) + c(rep(1,90), rep(-1,20), rep(1,90))
lambda <- 4
results <- narrowest_over_threshold(y, lambda, N=500)
print(results$results)

}
