% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_psi_change_in_mean.R
\name{not_psi_change_in_mean}
\alias{not_psi_change_in_mean}
\title{Post-selection inference for narrowest over threshold change in mean model.}
\usage{
not_psi_change_in_mean(
  y,
  lambda,
  results = NULL,
  max_cps = 1,
  N = 1000,
  rand_ints = NULL
)
}
\arguments{
\item{y}{Numeric vector of data.}

\item{lambda}{Threshold for determining changepoint candidates.}

\item{results}{Output of narrowest over threshold algorithm (optional).}

\item{max_cps}{Maximum number of changepoints for the algorithm to detect; defaults to 1; ignored if results
specified.}

\item{N}{Number of random intervals; defaults to 1000; ignored if results or rand_ints specified.}

\item{rand_ints}{Random intervals for narrowest over threshold algorithm; ignored if results specified.}
}
\value{
List of results ...
}
\description{
Calculates p-value for first detected changepoint, when the narrowest over threshold algorithm
is applied to the change in mean model.
}
\examples{
# ...

}
