% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wild_binary_segmentation.R
\name{wild_binary_segmentation}
\alias{wild_binary_segmentation}
\title{Wild Binary Segmentation}
\usage{
wild_binary_segmentation(
  y,
  num_rand_samples = 1000,
  random_samples = NULL,
  threshold = NULL,
  maxiter = NULL,
  seeded = FALSE,
  decay = sqrt(2)
)
}
\arguments{
\item{y}{Numeric vector of data.}

\item{num_rand_samples}{Number of random intervals to use. Ignored if \code{random_samples} is supplied.}

\item{random_samples}{N x 2 matrix of random intervals.}

\item{threshold}{Minimum threshold for determining changepoint candidates.}

\item{maxiter}{Maximum number of changepoints to find.}

\item{seeded}{Logical; if \code{TRUE}, seeded binary segmentation is used.}

\item{decay}{Decay parameter for seeded binary segmentation. Only used if \code{seeded = TRUE}.}
}
\value{
A list
}
\description{
Implements wild binary segmentation algorithm for changepoint detection.
}
\examples{
set.seed(100)
y <- rnorm(100) + c(rep(1,40), rep(0,10), rep(1,50))
results <- wild_binary_segmentation(y, num_rand_samples=500)
print(results$results)

}
