% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_segmentation_psi.R
\name{binary_segmentation_psi}
\alias{binary_segmentation_psi}
\title{Post-selection inference for binary segmentation}
\usage{
binary_segmentation_psi(
  y,
  results = NULL,
  nus = NULL,
  threshold = NULL,
  maxiter = NULL,
  eps0 = 0.01,
  sigma2 = NULL,
  h = NULL,
  first_cp_only = FALSE,
  num_pvals = NULL
)
}
\arguments{
\item{y}{Numeric vector of data.}

\item{results}{Output of binary_segmentation function; can be NULL.}

\item{nus}{List of nu's for each detected changepoint. See details.}

\item{threshold}{Numeric; minimum threshold for CUSUM statistic for changepoint detection. Ignored if results specified.}

\item{maxiter}{Number of changepoints to find. Ignored if results specified; otherwise defaults to \code{length(y) - 1}.}

\item{eps0}{Hyperparameter for calculating S. (Details.)}

\item{sigma2}{Variance of \code{y}. If unknown, it can be estimated within the function.}

\item{h}{Window size. See details.}

\item{first_cp_only}{Logical. If \code{TRUE}, condition on the fact that the changepoint of interest is in the model;
if \code{FALSE}, condition on all changepoints. Defaults to \code{TRUE} if \code{h} is supplied, and \code{FALSE} otherwise.}

\item{num_pvals}{Integer. Maximum number of p-values to calculate; defaults to \code{length(y) - 1}.}
}
\value{
A list containing:
\itemize{
\item \code{b} Numeric vector of changepoints
\item \code{p_value} Numeric vector of p-values
}
}
\description{
Calculate p-values associated with detected changepoints for binary segmentation;
uses method of Jewell et al.
}
\details{
Given a changepoint of interest \eqn{\tau_j}, there are two options for the null hypothesis:
\itemize{
\item There are no changepoints within a window size \eqn{h} of \eqn{\tau_j}. If \code{nus = NULL} but \code{h} is supplied,
then the value of \eqn{\nu} for each changepoint will be calculated using this assumption.
\item There are no other changepoints between \eqn{\tau_{j-1}} and \eqn{\tau_{j+1}}. If \eqn{\tau_j} is the first
changepoint, then \eqn{\tau_{j-1}} is taken to be 0; if it is the last changepoint, \eqn{\tau_{j+1}} is taken to be
\code{length(y)}. If \code{nus = NULL} and \code{h = NULL}, then the value of \eqn{\nu} for each changepoint will be calculated
using this assumption.
}
Alternatively \code{nus} can be specified manually.
}
\examples{
set.seed(100)
y <- rnorm(200) + c(rep(1,50), rep(0,100), rep(1,50))
results <- binary_segmentation(y)
pvals <- binary_segmentation_psi(y, results, sigma2=1)
print(pvals)

}
