% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_interval.R
\name{calculate_interval}
\alias{calculate_interval}
\title{Calculate interval}
\usage{
calculate_interval(
  y,
  method,
  results,
  nu,
  nu2 = NULL,
  nuTy = NULL,
  n.cp = NULL
)
}
\arguments{
\item{y}{Numeric vector of data}

\item{method}{Character string; \code{"bs"} for binary segmentation; \code{"wbs"} for wild binary segmentation;
\code{"not"} for narrowest over threshold.}

\item{results}{...}

\item{nu}{...}

\item{nu2}{value of ||nu||_2^2}

\item{nuTy}{value of nu^T y}

\item{n.cp}{max. number of changepoints to detect of binary segmentation algorithm}

\item{threshold}{minimum changepoint threshold used in binary segmentation algorithm}
}
\value{
A 2-dimensional vector
}
\description{
...
}
\examples{
x <- 0

}
