% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_interval_bs.R
\name{calculate_interval_bs}
\alias{calculate_interval_bs}
\title{Calculate interval for binary segmentation}
\usage{
calculate_interval_bs(
  y,
  nu,
  b,
  d,
  nu2 = NULL,
  nuTy = NULL,
  threshold = NULL,
  n.cp = NULL
)
}
\arguments{
\item{y}{Numeric vector of data.}

\item{nu}{...}

\item{b}{Vector of changepoints detected by binary segmentation algorithm.}

\item{d}{Directions of changepoints detected by binary segmentation algorithm. Vector whose entries are all equal to \code{1} or \code{-1}.}

\item{nu2}{Value of the squared 2-norm of \code{nu}.}

\item{threshold}{Minimum changepoint threshold used in binary segmentation algorithm.}

\item{n.cp}{Maximum number of changepoints to detect of binary segmentation algorithm.}
}
\value{
A 2-dimensional vector
}
\description{
Find values of phi which satisfy the required inequalities so that the binary segmentation algorithm returns (b, d);
Option to only consider part of (b, d), e.g. if we just want b1 & d1 to be specified, & don't care about
later values
}
\details{
(possibly outdated)
At present this ignores the threshold and just matches the number of CPs found (currently updating)
If threshold is also supplied, this will also be used.
Otherwise, if threshold is supplied (but not n.cp), the interval returned will be that for which the complete
b and d match those given, when BS is given this threshold. If threshold is not supplied, it will be assumed
that BS is run with a fixed number of iterations. If both n.cp and threshold are supplied, it will be assumed that
n.cp is the maximum number of iterations, but the threshold will also be used as a minimum for C(t).
}
\examples{
# to do

}
