% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_interval_bs.R
\name{calculate_interval_bs}
\alias{calculate_interval_bs}
\title{Calculate interval for binary segmentation}
\usage{
calculate_interval_bs(
  y,
  nu,
  b,
  d,
  nu2 = NULL,
  nuTy = NULL,
  threshold = NULL,
  n.cp = NULL
)
}
\arguments{
\item{y}{Numeric vector of data.}

\item{nu}{...}

\item{b}{Vector of changepoints detected by binary segmentation algorithm.}

\item{d}{Directions of changepoints detected by binary segmentation algorithm. Vector whose entries are all equal to \code{1}
or \code{-1}.}

\item{nu2}{Value of \eqn{||\nu||_2^2}; optional.}

\item{nuTy}{Value of \eqn{\nu^T y}; optional.}

\item{threshold}{Threshold used in binary segmentation algorithm.}

\item{n.cp}{Maximum number of changepoints to detect of binary segmentation algorithm.}
}
\value{
A 2-dimensional vector
}
\description{
Find values of phi which satisfy the required inequalities so that the binary segmentation algorithm returns (b, d);
Option to only consider part of (b, d), e.g. if we just want b1 & d1 to be specified, & don't care about
later values
}
\examples{
# to do

}
