% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pvals.R
\name{calculate_pvals}
\alias{calculate_pvals}
\title{Calculate p-values for change in mean model}
\usage{
calculate_pvals(
  y,
  method = "bs",
  results = NULL,
  N = 100,
  nus = NULL,
  threshold = NULL,
  maxiter = NULL,
  h = 2,
  sigma2 = 1,
  eps0 = 0.01,
  include_original = FALSE,
  num_pvals = NULL,
  random_samples = NULL,
  num_rand_samples = NULL,
  seeded = FALSE,
  decay = NULL
)
}
\arguments{
\item{y}{Numeric vector of data}

\item{method}{Method used to generate changepoints. Options: "bs" for binary segmentation,
"wbs" for wild or seeded binary segmentation, "not" for narrowest over threshold.}

\item{results}{Output of changepoint algorithm.}

\item{N}{Positive integer. Number of random psi's to generate.}

\item{nus}{List of nu's for each changepoint.}

\item{threshold}{Scalar. Minimum threshold for changepoint detection. Ignored if results is specified.}

\item{maxiter}{Positive integer. Maximum number of changepoints to find in changepoint algorithm. Ignored if
results is specified}

\item{h}{Positive integer (>=2) or NULL. Window size. If NULL then the changepoints either side of the
changepoint of interest are used to define the window.}

\item{sigma2}{Scalar. Variance of y.}

\item{eps0}{Scalar. Hyperparameter for calculating S.}

\item{include_original}{Logical. Whether to include the psi value corresponding to the observed data in place of
one of the ranom samples.}

\item{num_pvals}{Integer. Number of changepoints to calculate p-values for.}

\item{random_samples}{Matrix containing random intervals. For wild binary segmentation and narrowest over
threshold only.}

\item{num_rand_samples}{Number of random intervals to use for WBS/NOT. Ignored if random_samples is specified,
method = "bs", or seeded = TRUE.}

\item{seeded}{Logical. If TRUE, use seeded binary segmentation rather than wild binary segmentation. Only used if
method = "wbs".}

\item{decay}{Decay parameter for seeded binary segmentation. Only used if method = "wbs" and seeded = TRUE.}
}
\value{
A vector of p-values
}
\description{
Calculate p-values for change in mean model
}
\examples{
# to do

}
