% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_S_all_methods.R
\name{calculate_S_all_methods}
\alias{calculate_S_all_methods}
\title{Calculate S}
\usage{
calculate_S_all_methods(
  y,
  nu,
  results = NULL,
  b = NULL,
  d = NULL,
  nu2 = NULL,
  nuTy = NULL,
  threshold = NULL,
  maxiter = NULL,
  eps0 = 0.01,
  first_cp_only = FALSE,
  method = "bs",
  rand_ints = NULL,
  seeded = FALSE,
  decay = NULL
)
}
\arguments{
\item{y}{Numeric vector of data.}

\item{nu}{...}

\item{results}{Output of changepoint algorithm.}

\item{b}{Vector of changepoints; ignored if results specified.}

\item{d}{Vector containing directions of changepoints (+1 or -1); ignored if results specified.}

\item{nu2}{Scalar; value of ||nu||_2^2; optional.}

\item{nuTy}{Value of nu^T y.}

\item{threshold}{Threshold for detecting changepoints; ignored if results specified.}

\item{maxiter}{Maximum number of changepoints to detect; ignored if results specified.}

\item{eps0}{...}

\item{first_cp_only}{Logical. If \code{TRUE}, condition on the fact that the changepoint of interest is in the model;
if \code{FALSE}, condition on all changepoints. Defaults to \code{TRUE} if \code{h} is supplied, and \code{FALSE} otherwise.}

\item{method}{One of \code{"bs"} (binary segmentation), \code{"wbs"} (wild binary segmentation), or \code{"not"} (narrowest
over threshold). Defaults to \code{"bs"}.}

\item{rand_ints}{Matrix containing random intervals for changepoint algorithm. Ignored if results specified
or \code{method = "bs"}.}

\item{seeded}{Logical. For \code{method = "wbs"} only, whether to use seeded binary segmentation.}

\item{decay}{Decay parameter for seeded binary segmentation. Only used if \code{method = "wbs"} and \code{seeded = TRUE}.}

\item{num_pvals}{Integer. Maximum number of p-values to calculate; defaults to \code{length(y) - 1}.}
}
\value{
A dataframe containing intervals with the changepoints obtained when \eqn{\phi} is in each interval.
}
\description{
A function for calculating the set of interest S.
}
\details{
...
}
\examples{
# ...
}
