% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusum.R
\name{cusum}
\alias{cusum}
\title{Calculate CUSUM statistics for a vector}
\usage{
cusum(y, s = 1, e = length(y), return_full = FALSE)
}
\arguments{
\item{y}{A numeric vector.}

\item{s}{Starting index.}

\item{e}{Ending index.}

\item{return_full}{Whether to return full vector.}
}
\value{
A numeric vector.
}
\description{
Calculate CUSUM statistics for a subset of indices of y.
}
\examples{
y <- c(1,3,2,5,3)
cusum(y)

}
