% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusum_phi_vec.R
\name{cusum_phi_vec}
\alias{cusum_phi_vec}
\title{Calculate vector of CUSUM statistics for data y, in terms of phi}
\usage{
cusum_phi_vec(y, nu, nu2 = NULL, nuTy = NULL, s = 1, e = length(y))
}
\arguments{
\item{y}{Vector of data}

\item{nu}{nu}

\item{nu2}{Value of ||nu||_2^2}

\item{nuTy}{Value of nu^T y}

\item{s}{Starting point for calculating CUSUM statistics, defaults to 1}

\item{e}{Ending point for calculating CUSUM statistics, degaults to n}
}
\value{
Returns n x 2 matrix s.t. cusum_phi_vec(y)^T (1,phi) = cusm(y'(phi))
}
\description{
Used inside binary_segmentation_psi (etc.)
}
\examples{
set.seed(100)
y <- rnorm(100) + c(rep(1,40), rep(-1,60))
results <- binary_segmentation(y)
b <- results$changepoints[1]
nu <- c(rep(1/b, b), rep(-1/(100-b), 100-b))
cusum_phi_vec(y, nu)

}
