% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cps.R
\name{find_cps}
\alias{find_cps}
\title{Title}
\usage{
find_cps(
  y,
  method,
  threshold = NULL,
  maxiter = NULL,
  num_rand_ints = NULL,
  rand_ints = NULL,
  seeded = FALSE,
  decay = NULL
)
}
\arguments{
\item{y}{vector of data}

\item{method}{"bs" = binary segmenation; "wbs" = wild binary segmentation; "not" = narrowest over threshold}

\item{threshold}{changepoint detection threshold for CUSUM statistic}

\item{maxiter}{max. number of changepoints to detect}

\item{num_rand_ints}{number of random intervals (ignored if rand_ints specified, or if method = "bs")}

\item{rand_ints}{random intervals for wild binary segmentation or narrowest over threshold (ignored if method = "bs")}

\item{seeded}{if TRUE and method = "wbs", then seeded binary segmentation is implemented}

\item{decay}{decay parameter for seeded binary segmentation; only used if method = "wbs" and seeded = TRUE}
}
\value{
...
}
\description{
Implements binary segmentation, wild binary segmentation, narrowest over threshold for change in mean model.
}
\examples{
# ...

}
