% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_random_intervals.R
\name{generate_random_intervals}
\alias{generate_random_intervals}
\title{Generate random intervals}
\usage{
generate_random_intervals(n, N, min_width = 2)
}
\arguments{
\item{n}{Integer; upper limit}

\item{N}{Integer; number of intervals to generate}

\item{min_width}{Integer; minimum width of interval allowed}
}
\value{
An N x 2 matrix of integers
}
\description{
Generate random intervals within the range (1,n). For use in wild binary segmentation and narrowest over threshold algorithms.
}
\details{
If the number of intervals supplied is greater than the total number of possible unique intervals contained in \link{1,n} with width at least
\code{min_width}, then the output will be a list of all possible intervals contained in \link{1,n}.
}
\examples{
generate_random_intervals(100, 10, 2)

}
