% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l0_segmentation_psi.R
\name{l0_segmentation_psi}
\alias{l0_segmentation_psi}
\title{Post-selection inference for L0 segmentation}
\usage{
l0_segmentation_psi(
  y,
  lambda,
  N,
  h,
  sigma2 = 1,
  sig = 4,
  include_original = TRUE,
  num_pvals = NULL
)
}
\arguments{
\item{y}{Numeric vector of data}

\item{lambda}{Threshold parameter}

\item{N}{Number of samples of psi to take}

\item{h}{Window size}

\item{sigma2}{Variance of y}

\item{sig}{Tuning parameter}

\item{include_original}{Logical; whether to include observed value as psi in place as one of the random samples;
defaults to \code{TRUE}}

\item{num_pvals}{Maximum number of p-values to calculate}
}
\value{
...
}
\description{
Post-selection inference for L0 segmentation. This uses the functions \code{changepoint_estimates} and
\code{changepoint_inference} from the package \code{ChangepointInference}.
}
\examples{
...

}
