% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narrowest_over_threshold.R
\name{narrowest_over_threshold}
\alias{narrowest_over_threshold}
\title{Narrowest over threshold for change in mean.}
\usage{
narrowest_over_threshold(y, lambda, rand_ints = NULL, N = 1000, max_cps = NULL)
}
\arguments{
\item{y}{A numeric vector of data.}

\item{lambda}{Threshold; a scalar.}

\item{rand_ints}{Matrix containing random intervals; optional.}

\item{N}{Number of random intervals; defaults to 1000.}

\item{max_cps}{Maximum number of changepoints to return.}
}
\value{
...
}
\description{
Narrowest over threshold for change in mean.
}
\examples{
y <- rnorm(200) + c(rep(1,90), rep(-1,20), rep(1,90))
lambda <- 4
results <- narrowest_over_threshold(y, lambda, N=500)
print(results$results)

}
