% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_phi.R
\name{y_phi}
\alias{y_phi}
\title{Calculate y'(phi)}
\usage{
y_phi(y, nu, phi, nu2 = NULL, nuTy = NULL)
}
\arguments{
\item{y}{Numeric data vector.}

\item{nu}{Numeric vector of same length as \code{y}.}

\item{phi}{Numeric.}

\item{nu2}{Value of \eqn{||\nu||_2^2}; optional.}

\item{nuTy}{Value of \eqn{\nu^T y}; optional.}
}
\value{
Numeric vector.
}
\description{
Calculate \eqn{y'(\phi) = y - (||\nu||_2^2)^{-1} \nu \nu^T y + (||\nu||_2^2)^{-1} \nu \phi}.
}
\examples{
y <- rnorm(20) + c(rep(1, 10), rep(-1, 10))
nu <- c(rep(0, 5), rep(1/5, 5), rep(-1/5, 5), rep(0, 5))
phi <- 2
y_phi(y, nu, phi)

}
